# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API",
        "description": "API to manage all endpoints",
        "version": "1.0.0"
    },
    "paths": {
        "/api/category": {
            "get": {
                "tags": [
                    "categories"
                ],
                "summary": "Get all categories",
                "description": "Retrieve a list of all categories",
                "operationId": "54ade06619716ab9ebfa4797702cfa08",
                "responses": {
                    "200": {
                        "description": "List of categories"
                    },
                    "404": {
                        "description": "No category found"
                    }
                }
            },
            "post": {
                "tags": [
                    "Categories"
                ],
                "summary": "Create a new category",
                "description": "Create a new category with provided name and age",
                "operationId": "4b1f63f71e949f4fca52ca76229a33a4",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "description"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string"
                                    },
                                    "description": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Category created"
                    },
                    "400": {
                        "description": "Bad request"
                    }
                }
            }
        },
        "/api/category/{id}": {
            "get": {
                "tags": [
                    "categories"
                ],
                "summary": "Get a category by ID",
                "description": "Retrieve a category by its ID",
                "operationId": "c2704f9826d06e00fc30d7729f17020d",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the category to retrieve",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Category found"
                    },
                    "404": {
                        "description": "Category not found"
                    }
                }
            },
            "put": {
                "tags": [
                    "categories"
                ],
                "summary": "Update a category",
                "description": "Update the details of a category",
                "operationId": "0189c280308cbd436022df4ee8198af5",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the category to update",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "description"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string"
                                    },
                                    "description": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "category updated"
                    },
                    "400": {
                        "description": "Bad request"
                    },
                    "404": {
                        "description": "category not found"
                    }
                }
            },
            "delete": {
                "tags": [
                    "categories"
                ],
                "summary": "Delete a category",
                "description": "Delete a category by its ID",
                "operationId": "a65fe84035fb7e9096ca0c4bafa91cdc",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the category to delete",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Category deleted"
                    },
                    "404": {
                        "description": "Category not found"
                    }
                }
            }
        },
        "/api/event": {
            "get": {
                "tags": [
                    "events"
                ],
                "summary": "Get all events",
                "description": "Retrieve a list of all events",
                "operationId": "256922707f030d4536b2cc4beaced12e",
                "responses": {
                    "200": {
                        "description": "List of events"
                    },
                    "404": {
                        "description": "No event found"
                    }
                }
            },
            "post": {
                "tags": [
                    "Events"
                ],
                "summary": "Create a new event",
                "description": "Create a new event with provided name and age",
                "operationId": "c921732b897774f6a6f45cc1ae813313",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "title",
                                    "description",
                                    "content",
                                    "event_date",
                                    "sport_type_id",
                                    "city_id",
                                    "image"
                                ],
                                "properties": {
                                    "title": {
                                        "type": "string"
                                    },
                                    "description": {
                                        "type": "string"
                                    },
                                    "content": {
                                        "type": "string"
                                    },
                                    "event_date": {
                                        "type": "string"
                                    },
                                    "sport_type_id": {
                                        "type": "integer"
                                    },
                                    "city_id": {
                                        "type": "integer"
                                    },
                                    "image": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Student created"
                    },
                    "400": {
                        "description": "Bad request"
                    }
                }
            }
        },
        "/api/event/{id}": {
            "get": {
                "tags": [
                    "events"
                ],
                "summary": "Get a event by ID",
                "description": "Retrieve a event by its ID",
                "operationId": "f1768354753eb16fe3998c397ee9e89a",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the student to retrieve",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Event found"
                    },
                    "404": {
                        "description": "Event not found"
                    }
                }
            },
            "put": {
                "tags": [
                    "events"
                ],
                "summary": "Update a event",
                "description": "Update the details of a event",
                "operationId": "ab9c6818e0145d03c8c4bd3c2e6d3a3c",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the event to update",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "title",
                                    "description",
                                    "content",
                                    "event_date",
                                    "sport_type_id",
                                    "city_id",
                                    "image"
                                ],
                                "properties": {
                                    "title": {
                                        "type": "string"
                                    },
                                    "description": {
                                        "type": "string"
                                    },
                                    "content": {
                                        "type": "string"
                                    },
                                    "event_date": {
                                        "type": "string"
                                    },
                                    "sport_type_id": {
                                        "type": "integer"
                                    },
                                    "city_id": {
                                        "type": "integer"
                                    },
                                    "image": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "event updated"
                    },
                    "400": {
                        "description": "Bad request"
                    },
                    "404": {
                        "description": "event not found"
                    }
                }
            },
            "delete": {
                "tags": [
                    "events"
                ],
                "summary": "Delete a event",
                "description": "Delete a event by its ID",
                "operationId": "97e1d312b1646b375b2dfad9ce146ea9",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the event to delete",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Event deleted"
                    },
                    "404": {
                        "description": "Event not found"
                    }
                }
            }
        }
    }
}